ChangeCustomerDataNegativeFlowTest.java was deleted, ChangeCustomerDataTest.java renamed from ChangeCustomerDataPositiveFlowTest.java
